Scoring Class Template
Just a basic outline of psuedocode, and the beginning phases of implementation(defined attributes, functions, but have not yet implemented the code which uses them) I have a rough idea on what to put. As well as a bit of testing to see what I can, or, can't do with my available options in approach.-Andre Also, I've set the loseBackground to offscreen, as well as disabled it for now, I'm not sure if it will be  nuisance or get in the way if fully integrated to main. For now, I'll keep working on it and try learning the property options available towards UI elements, so will just pull to my branch.-Andre.